fix: add explicit uint32_t casts to resolve MISRA-C 12.2 violations
Add explicit uint32_t casts for bit shift operations in CAN packet setter
functions to resolve MISRA-C 2012 rule 12.2 violations. This ensures the
shift operands are within the proper range for 32-bit unsigned integers.

- Cast val parameter to uint32_t before bit shifts in SET_EXTENDED
- Cast val parameter to uint32_t before bit shifts in SET_RETURNED
- Cast val parameter to uint32_t before bit shifts in SET_REJECTED
- Maintains identical functionality while being MISRA-compliant

+++ board/can_declarations.h
@@ -57,11 +57,11 @@ static inline void SET_ADDR(CANPacket_t *msg, uint32_t val) {
   msg->addr = (msg->addr & 0xE0000000U) | ((val) & 0x1FFFFFFFU);
 }
 static inline void SET_EXTENDED(CANPacket_t *msg, uint8_t val) {
-  msg->addr = (msg->addr & 0x7FFFFFFFU) | (((val) & 1U) << 31U);
+  msg->addr = (msg->addr & 0x7FFFFFFFU) | (((uint32_t)(val) & 1U) << 31U);
 }
 static inline void SET_RETURNED(CANPacket_t *msg, uint8_t val) {
-  msg->addr = (msg->addr & 0xBFFFFFFFU) | (((val) & 1U) << 30U);
+  msg->addr = (msg->addr & 0xBFFFFFFFU) | (((uint32_t)(val) & 1U) << 30U);
 }
 static inline void SET_REJECTED(CANPacket_t *msg, uint8_t val) {
-  msg->addr = (msg->addr & 0xDFFFFFFFU) | (((val) & 1U) << 29U);
+  msg->addr = (msg->addr & 0xDFFFFFFFU) | (((uint32_t)(val) & 1U) << 29U);
 }
